refactor(print-format-builder): move sanitize_html to utils.js
